# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "description": "전체 회원 PublicUser 목록. Bearer 없이 호출 가능. passwordHash/refreshTokens 제외. 기본은 게스트 제외.",
        "operationId": "UsersController_listUsers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": "1",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "최대 100",
            "schema": {
              "example": "50",
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "nickname / email 부분 검색",
            "schema": {
              "example": "윤지",
              "type": "string"
            }
          },
          {
            "name": "includeGuests",
            "required": false,
            "in": "query",
            "description": "true면 게스트 초대 유저 포함",
            "schema": {
              "example": "false",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListPageDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "회원가입한 유저 목록 (인증 불필요)",
        "tags": [
          "유저"
        ]
      }
    },
    "/api/v1/users/me": {
      "get": {
        "description": "온보딩·퀴즈·성향 축 등 본인 필드 전부 + stats. passwordHash/refreshTokens는 제외.",
        "operationId": "UsersController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "내 전체 프로필 + 여행 통계",
        "tags": [
          "유저"
        ]
      }
    },
    "/api/v1/users/me/travel-type": {
      "get": {
        "operationId": "UsersController_getTravelType",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TravelTypeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "두리 테스트 결과 (TravelType)",
        "tags": [
          "유저"
        ]
      }
    },
    "/api/v1/users/me/trips-summary": {
      "get": {
        "operationId": "UsersController_getTripsSummary",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripsSummaryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "ongoing / completed 여행 수",
        "tags": [
          "유저"
        ]
      }
    },
    "/api/v1/users/me/profile": {
      "patch": {
        "operationId": "UsersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "프로필 수정",
        "tags": [
          "유저"
        ]
      }
    },
    "/api/v1/users/me/onboarding-complete": {
      "patch": {
        "operationId": "UsersController_completeOnboarding",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "온보딩 완료 처리",
        "tags": [
          "유저"
        ]
      }
    },
...
  },
  "info": {
    "title": "Tourmate API",
    "description": "TripMatch / Tourmate backend REST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TravelTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "감성 탐험가"
          },
          "description": {
            "type": "string",
            "example": "분위기 좋은 곳을 찾아다니는 타입"
          },
          "tags": {
            "example": [
              "감성",
              "카페",
              "사진"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "warning": {
            "type": "string",
            "example": "동선이 길어질 수 있어요"
          },
          "emoji": {
            "type": "string",
            "example": "📷"
          }
        },
        "required": [
          "name",
          "description",
          "tags",
          "warning",
          "emoji"
        ]
      },
      "PublicUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "665abc123def456789012345"
          },
          "email": {
            "type": "string",
            "example": "test@example.com"
          },
          "nickname": {
            "type": "string",
            "example": "윤지"
          },
          "profileImageUrl": {
            "type": "object",
            "example": "https://example.com/me.jpg",
            "nullable": true
          },
          "travelType": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/TravelTypeDto"
              }
            ]
          },
          "onboardingCompleted": {
            "type": "boolean",
            "example": false
          },
          "isGuest": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "id",
          "nickname",
          "onboardingCompleted",
          "isGuest"
        ]
      },
...
      "UserListPageDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicUserDto"
            }
          },
          "meta": {
            "type": "object",
            "example": {
              "total": 42,
              "page": 1,
              "limit": 50
            }
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "PersonalityAxesDto": {
        "type": "object",
        "properties": {
          "scheduleDensity": {
            "type": "number",
            "example": 72
          },
          "landmarkNecessity": {
            "type": "number",
            "example": 55
          },
          "localInterest": {
            "type": "number",
            "example": 40
          },
          "challenging": {
            "type": "number",
            "example": 80
          }
        },
        "required": [
          "scheduleDensity",
          "landmarkNecessity",
          "localInterest",
          "challenging"
        ]
      },
      "MeStatsDto": {
        "type": "object",
        "properties": {
          "ongoingTrips": {
            "type": "number",
            "example": 2
          },
          "completedTrips": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "ongoingTrips",
          "completedTrips"
        ]
      },
      "MeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "665abc123def456789012345"
          },
          "email": {
            "type": "string",
            "example": "test@example.com"
          },
          "nickname": {
            "type": "string",
            "example": "윤지"
          },
          "profileImageUrl": {
            "type": "object",
            "example": "https://example.com/me.jpg",
            "nullable": true
          },
          "travelType": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/TravelTypeDto"
              }
            ]
          },
          "onboardingCompleted": {
            "type": "boolean",
            "example": false
          },
          "isGuest": {
            "type": "boolean",
            "example": false
          },
          "oauthProvider": {
            "type": "object",
            "example": "kakao",
            "nullable": true
          },
          "quizPreferences": {
            "type": "object",
            "nullable": true,
            "description": "두리 테스트 preference 캐시",
            "example": {
              "transport": "car",
              "budget": "mid"
            }
          },
          "personalityAxes": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PersonalityAxesDto"
              }
            ]
          },
          "hasLicense": {
            "type": "object",
            "example": true,
            "nullable": true
          },
          "hasCar": {
            "type": "object",
            "example": false,
            "nullable": true
          },
          "mobilityConstraints": {
            "example": [
              "STAIRS"
            ],
            "description": "STAIRS | STEEP_SLOPE | LONG_WALK",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "birthYear": {
            "type": "object",
            "example": 1999,
            "nullable": true
          },
          "interestTags": {
            "example": [
              "카페",
              "자연"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "object",
            "nullable": true
          },
          "updatedAt": {
            "type": "object",
            "nullable": true
          },
          "stats": {
            "$ref": "#/components/schemas/MeStatsDto"
          }
        },
        "required": [
          "id",
          "nickname",
          "onboardingCompleted",
          "isGuest",
          "mobilityConstraints",
          "interestTags",
          "stats"
        ]
      },
      "TripsSummaryDto": {
        "type": "object",
        "properties": {
          "ongoing": {
            "type": "number",
            "example": 2
          },
          "completed": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "ongoing",
          "completed"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string",
            "example": "윤지"
          },
          "profileImageUrl": {
            "type": "string",
            "example": "https://example.com/profile.jpg"
          }
        }
      },
...
    }
  }
}